src/main/java/com/springmvc/springmvc/security/SecurityUtil.java
package com.springmvc.springmvc.security;

import org.springframework.security.authentication.AnonymousAuthenticationToken;
import org.springframework.security.core.Authentication;
import org.springframework.security.core.context.SecurityContextHolder;

public final class SecurityUtil {
    
    public static String getSessionUser() {
        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
        if (!(authentication instanceof AnonymousAuthenticationToken) && authentication!=null) {
            String currentUsername = authentication.getName();
            return currentUsername;
        }
        return null;
    }
}
